Refactor validation utils into smaller modules
